split test files by driver
